+++ .travis.yml
@@ -1,14 +1,16 @@
 dist: xenial
 language: python
 python:
-- '2.7'
+- '3.4'
 - '3.5'
 - '3.6'
 - '3.7'
+- '3.8'
+- '3.9'
 install:
 - pip install -U setuptools pip -r build-requirements.txt
 script:
-- pytest --doctest-modules --cov=anybadge --cov-report html:htmlcov anybadge.py tests
+- pytest --doctest-modules --cov=anybadge --cov-report html:htmlcov anybadge tests
 before_deploy:
 - sed -i "s/^version = .*/version = __version__ = \"$TRAVIS_TAG\"/" anybadge.py
 deploy:
@@ -21,4 +23,4 @@ deploy:
   on:
     tags: true
     all_branches: true
-    python: '3.7'
+    python: '3.9'
