+++ anybadge/colors.py
@@ -1,0 +1,27 @@
+# Create a dictionary of colors to make selections
+# easier.
+from enum import Enum
+
+
+class Color(Enum):
+    WHITE = '#FFFFFF'
+    SILVER = '#C0C0C0'
+    GRAY = '#808080'
+    BLACK = '#000000'
+    RED = '#e05d44'
+    BRIGHT_RED = '#FF0000'
+    MAROON = '#800000'
+    OLIVE = '#808000'
+    LIME = '#00FF00'
+    BRIGHT_YELLOW = '#FFFF00'
+    YELLOW = '#dfb317'
+    GREEN = '#4c1'
+    YELLOW_GREEN = '#a4a61d'
+    AQUA = '#00FFFF'
+    TEAL = '#008080'
+    BLUE = '#0000FF'
+    NAVY = '#000080'
+    FUCHSIA = '#FF00FF'
+    PURPLE = '#800080'
+    ORANGE = '#fe7d37'
+    LIGHT_GREY = '#9f9f9f'
