+++ anybadge/config.py
@@ -1,0 +1,22 @@
+# Set some defaults
+DEFAULT_FONT = 'DejaVu Sans,Verdana,Geneva,sans-serif'
+DEFAULT_FONT_SIZE = 11
+NUM_PADDING_CHARS = 0.5
+DEFAULT_COLOR = '#4c1'
+DEFAULT_TEXT_COLOR = '#fff'
+MASK_ID_PREFIX = 'anybadge_'
+
+# Dictionary for looking up approx pixel widths of
+# supported fonts and font sizes.
+FONT_WIDTHS = {
+    'DejaVu Sans,Verdana,Geneva,sans-serif': {
+        10: 9,
+        11: 10,
+        12: 11,
+    },
+    'Arial, Helvetica, sans-serif': {
+        11: 8,
+    },
+}
+
+
