+++ anybadge/exceptions.py
@@ -1,0 +1,3 @@
+
+class UnknownBadgeTemplate(Exception):
+    """The badge template is unknown."""
