+++ anybadge/server/config.py
@@ -1,0 +1,9 @@
+import logging
+
+
+DEFAULT_SERVER_PORT = 8000
+DEFAULT_SERVER_LISTEN_ADDRESS = 'localhost'
+DEFAULT_LOGGING_LEVEL = logging.INFO
+
+SERVER_PORT = DEFAULT_SERVER_PORT
+SERVER_LISTEN_ADDRESS = DEFAULT_SERVER_LISTEN_ADDRESS
